% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 1 .R
\name{count_pts_LeBron}
\alias{count_pts_LeBron}
\title{LeBron James Points function}
\usage{
count_pts_LeBron(year)
}
\arguments{
\item{year}{the year you want to know about}
}
\description{
This function allows you to know how many points LeBron James scored for any given year.
}
\examples{
count_pts_LeBron()
}
\keyword{James}
\keyword{LeBron}
