% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list.R
\name{DEG_list}
\alias{DEG_list}
\title{Perform Differential Expression Analysis}
\usage{
DEG_list(
  Phenotype_data,
  Expression_data,
  Annotation_data,
  p_cutoff = 0.05,
  fc_cutoff = 1
)
}
\arguments{
\item{Phenotype_data}{pheno type data}

\item{Expression_data}{expression matrix}

\item{Annotation_data}{annotation data}
}
\value{
a list containing differentially expressed genes and statistical results
}
\description{
This function performs differential expression analysis using limma.
}
\examples{
result <- DEG_list(Phenotype_data, Expression_data, Annotation_data)
}
