% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGSE.R
\name{DGSE}
\alias{DGSE}
\title{Perform Disease Gene Set Enrichment Analysis (GSEA) using identified DEGs}
\usage{
DGSE(DEG_results, p_value_cutoff = 0.05)
}
\arguments{
\item{DEG_results}{Differential expression results data frame with columns 'logFC', 'pvalue', and 'adj.P.Val'.}

\item{p_value_cutoff}{P-value cutoff for selecting DEGs.}
}
\value{
the result table of disease gene set enrichment analysis
}
\description{
This function performs Disease GSEA on identified DEGs.
}
\examples{
DGSE(DEG_results, p.cutoff=0.05)
DEG_results <- ...  # Your differential expression results data frame
DGSE(DEG_results, p_value_cutoff = 0.2)

}
