% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{Perform Gene Set Enrichment Analysis (GSEA) on identified DEGs}
\usage{
GSEA(DEG_results, p_value_cutoff = 0.05)
}
\arguments{
\item{DEG_results}{Differential expression results data frame with columns 'logFC', 'pvalue', and 'adj.P.Val'.}

\item{p_value_cutoff}{P-value cutoff for selecting DEGs.}
}
\value{
A list of GSEA results.
}
\description{
Perform Gene Set Enrichment Analysis (GSEA) on identified DEGs
}
\examples{
DEG_results <- ...  # Your differential expression results data frame
GSEA(DEG_results, p_value_cutoff = 0.05)
}
