\name{Homework5-package}
\alias{Homework5-package}
\alias{Homework5}
\docType{package}
\title{Homework5
\packageTitle{Homework5}
}
\description{
\packageDescription{Homework5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Homework5}
\packageIndices{Homework5}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Homework5}

Maintainer: \packageMaintainer{Homework5}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
