% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_GEO.R
\name{import_GEO}
\alias{import_GEO}
\title{Importing GEO object and files}
\usage{
import_GEO(GEO_ID = "GSE33126")
}
\arguments{
\item{GEO_ID}{GEO ID for the dataset

\itemize{
  \item \code{Phenotype_data}: Phenotype data.
  \item \code{Expression_data}: Expression data.
  \item \code{Annotation_data}: Annotation data.
}}
}
\value{
A list of Expression, Phenotype and Annotation data
}
\description{
Importing GEO object and files
}
\examples{
my_id <- "GSE33126"
import_GEO(my_id)
}
