% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform_boxplot.R
\name{log_normalize_data}
\alias{log_normalize_data}
\title{Log-normalize the data}
\usage{
log_normalize_data(expr, boxplot = TRUE)
}
\arguments{
\item{expr}{Expression data matrix.}
}
\value{
Log-normalized expression data and boxplot.

the log normalized expression data and boxplot
}
\description{
Log-normalize the data
}
\examples{
log_normalized_data <- log_normalize_data(expression_data)
}
