% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Perform clustering and display heatmap with annotations}
\usage{
make_heatmap(Expression_data, Phenotype_data)
}
\arguments{
\item{Expression_data}{Expression data matrix.}

\item{Phenotype_data}{Sample information data frame with columns 'group' and 'patient'.}
}
\value{
A heatmap displaying clustering of samples with annotations.
}
\description{
This function performs clustering on the expression data and displays a heatmap with annotations.
}
\examples{
expression_data <- ...  # Your expression data
sample_info <- ...     # Your sample information data frame
make_heatmap(expression_data, sample_info)
}
