% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ma.R
\name{make_ma}
\alias{make_ma}
\title{Create a volcano plot with highlighted DEGs}
\usage{
make_ma(results, p.cutoff = 0.05, lfc.cutoff = 1)
}
\arguments{
\item{results}{Differential expression results data frame with columns 'logFC', 'pvalue', and 'adj.P.Val'.}

\item{p.cutoff}{p value cutoff}

\item{lfc.cutoff}{log fold change cutoff}
}
\value{
A MA plot with highlighted DEGs.
}
\description{
This function generates a volcano plot for differential gene expression results, highlighting DEGs.
}
\examples{
make_ma(full_results)
}
