% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pca.R
\name{make_pca}
\alias{make_pca}
\title{Create a scatter plot using Principal Component Analysis (PCA) vectors}
\usage{
make_pca(log_normalize_data, phenotype_data1)
}
\arguments{
\item{log_normalize_data}{Log transformed expression data matrix.}

\item{pheno}{Sample information data frame with columns 'group' and 'patient'.}
}
\value{
A scatter plot of the first two principal components.
}
\description{
This function performs PCA on the provided expression data and creates a scatter plot.
}
\examples{
make_pca <- pca_plot(expression_data, pheno)
}
