% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 3.R
\name{numeric_corr_matrix}
\alias{numeric_corr_matrix}
\title{Numeric Correlation Matrix Function}
\usage{
numeric_corr_matrix(year)
}
\arguments{
\item{year}{the year for which you want to know the correlation for the numeric values.}
}
\description{
This function allows you to find the correlation between all the numeric values for a given year.
}
\examples{
numeric_corr_matrix()
}
\keyword{correlation}
\keyword{matrix}
