% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function 2.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{the year for which you want to know the oldest player}
}
\description{
This function allows you to
}
\examples{
oldest_player()
}
\keyword{Oldest}
\keyword{player}
