% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_path.R
\name{reactome_path}
\alias{reactome_path}
\title{Perform Reactome pathway analysis}
\usage{
reactome_path(DEG_results, p_cutoff = 0.05)
}
\arguments{
\item{DEG_results}{A data frame containing DEGs with columns 'Entrez_Gene_ID', 'logFC', and 'pvalue'.}

\item{p_cutoff}{The p-value cutoff for enrichment analysis. Default is 0.05.}
}
\value{
A data frame with Reactome pathway enrichment results.
}
\description{
Perform Reactome pathway analysis
}
\examples{
DEG_results <- ...  # Your data frame with DEGs
result <- do_reactome_analysis(top_genes, p_value_cutoff = 0.05)
}
