% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square of a number}
\usage{
square(x)
}
\arguments{
\item{number}{double: a number given by the user}
}
\value{
The square of this number
}
\description{
Square of a number
}
\examples{
square(3)
}

