% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vis_network.R
\name{Vis_network}
\alias{Vis_network}
\title{Visualize Reactome pathway as a network}
\usage{
Vis_network(DEG_results, name)
}
\arguments{
\item{DEG_results}{DE analysis results}

\item{name}{name of the pathway(#Select one of the pathway descriptions from your over-represented Reactome Pathways)}
}
\value{
network plot of Reactome Pathway
}
\description{
Visualize Reactome pathway as a network
}
\examples{
Network(DEG_results, "Degradation of the extracellular matrix")
}
